Moving queue.h to the audio part
