+++ navit/audio/queue.h
